# src/main/java/com/boehringer/componentprovisioner/server/api/ProvisionerMessagesDefinitionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.boehringer.componentprovisioner.server.api;

import java.util.Map;
import com.boehringer.componentprovisioner.server.model.ProvisionerMessageDefinition;
import com.boehringer.componentprovisioner.server.model.RestErrorMessage;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
@Controller
@Tag(name = "ProvisionerMessagesDefinitions", description = "Provisioner standardized messages definitions")
public interface ProvisionerMessagesDefinitionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /catalog-items/{catalogItemId}/message-definitions/{id} : Get a message definition by catalogItemId and id.
     * Returns an standard message definition 
     *
     * @param catalogItemId id for the Catalog Item where Message is defined. (required)
     * @param id id for the MessageDefinition. (required)
     * @param requestBody  (required)
     * @return A single message definition. (status code 200)
     *         or Invalid client token on the request. (status code 401)
     *         or Insufficient permissions for the client to access the resource. (status code 403)
     *         or Catalog not found (status code 404)
     *         or Server error. (status code 500)
     */
    @Operation(
        operationId = "getMessageDefinitionByCatalogItemIdAndMessageId",
        summary = "Get a message definition by catalogItemId and id.",
        description = "Returns an standard message definition ",
        tags = { "ProvisionerMessagesDefinitions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A single message definition.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProvisionerMessageDefinition.class))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid client token on the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "403", description = "Insufficient permissions for the client to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "404", description = "Catalog not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "500", description = "Server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/catalog-items/{catalogItemId}/message-definitions/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ProvisionerMessageDefinition> getMessageDefinitionByCatalogItemIdAndMessageId(
        @Parameter(name = "catalogItemId", description = "id for the Catalog Item where Message is defined.", required = true, in = ParameterIn.PATH) @PathVariable("catalogItemId") String catalogItemId,
        @Parameter(name = "id", description = "id for the MessageDefinition.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "request_body", description = "", required = true) @Valid @RequestBody Map<String, String> requestBody
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"OPENSHIFT_CONNECTION_ERROR\", \"title\" : \"An error occurred while connecting to OpenShift\", \"message\" : \"Authorization error: please check your user credentials for deployment and try again later.\n\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /message-definitions/{id} : Get a message definition by id.
     * Returns an standard message definition 
     *
     * @param id id for the MessageDefinition. (required)
     * @param requestBody  (required)
     * @return A single message definition. (status code 200)
     *         or Invalid client token on the request. (status code 401)
     *         or Insufficient permissions for the client to access the resource. (status code 403)
     *         or Catalog not found (status code 404)
     *         or Server error. (status code 500)
     * @deprecated
     */
    @Deprecated
    @Operation(
        operationId = "getMessageDefinitionById",
        summary = "Get a message definition by id.",
        description = "Returns an standard message definition ",
        deprecated = true,
        tags = { "ProvisionerMessagesDefinitions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A single message definition.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProvisionerMessageDefinition.class))
            }),
            @ApiResponse(responseCode = "401", description = "Invalid client token on the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "403", description = "Insufficient permissions for the client to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "404", description = "Catalog not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            }),
            @ApiResponse(responseCode = "500", description = "Server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = RestErrorMessage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/message-definitions/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ProvisionerMessageDefinition> getMessageDefinitionById(
        @Parameter(name = "id", description = "id for the MessageDefinition.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "request_body", description = "", required = true) @Valid @RequestBody Map<String, String> requestBody
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : \"OPENSHIFT_CONNECTION_ERROR\", \"title\" : \"An error occurred while connecting to OpenShift\", \"message\" : \"Authorization error: please check your user credentials for deployment and try again later.\n\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}